repo_results_2/HumanEval/129/task_129.py
def minPath(grid, k):
	"""
	Given a grid with N rows and N columns (N >= 2) and a positive integer k, 
	each cell of the grid contains a value. Every integer in the range [1, N * N]
	inclusive appears exactly once on the cells of the grid.

	You have to find the minimum path of length k in the grid. You can start
	from any cell, and in each step you can move to any of the neighbor cells,
	in other words, you can go to cells which share an edge with you current
	cell.
	Please note that a path of length k means visiting exactly k cells (not
	necessarily distinct).
	You CANNOT go off the grid.
	A path A (of length k) is considered less than a path B (of length k) if
	after making the ordered lists of the values on the cells that A and B go
	through (let's call them lst_A and lst_B), lst_A is lexicographically less
	than lst_B, in other words, there exist an integer index i (1 <= i <= k)
	such that lst_A[i] < lst_B[i] and for any j (1 <= j < i) we have
	lst_A[j] = lst_B[j].
	It is guaranteed that the answer is unique.
	Return an ordered list of the values on the cells that the minimum path go through.

	Examples:

		Input: grid = [ [1,2,3], [4,5,6], [7,8,9]], k = 3
		Output: [1, 2, 1]

		Input: grid = [ [5,9,3], [4,1,6], [7,8,2]], k = 1
		Output: [1]
	"""

	# Flatten the grid into a 1D list
	flat_grid = [item for sublist in grid for item in sublist]

	# Find the minimum value in the grid
	min_val = min(flat_grid)

	# Find the position of the minimum value in the grid
	for i in range(len(grid)):
		for j in range(len(grid[i])):
			if grid[i][j] == min_val:
				min_pos = (i, j)

	# Initialize the path with the minimum value
	path = [min_val]

	# Find the minimum path of length k
	for _ in range(k - 1):
		min_neighbor_val = float('inf')
		min_neighbor_pos = None

		# Check the neighbors of the current cell
		for dx, dy in [(-1, 0), (1, 0), (0, -1), (0, 1)]:
			new_pos = (min_pos[0] + dx, min_pos[1] + dy)

			# If the new position is within the grid
			if 0 <= new_pos[0] < len(grid) and 0 <= new_pos[1] < len(grid[0]):
				# If the value of the new position is less than the current minimum neighbor value
				if grid[new_pos[0]][new_pos[1]] < min_neighbor_val:
					# Update the minimum neighbor value and its position
					min_neighbor_val = grid[new_pos[0]][new_pos[1]]
					min_neighbor_pos = new_pos

		# Add the minimum neighbor value to the path
		path.append(min_neighbor_val)

		# Update the current cell's position to the minimum neighbor's position
		min_pos = min_neighbor_pos

	# Return the path
	return path
